+++ Demo/Application/Base/BraintreeDemoSceneDelegate.h
@@ -1,0 +1,7 @@
+#import <UIKit/UIKit.h>
+
+@interface BraintreeDemoSceneDelegate : UIResponder <UIWindowSceneDelegate>
+
+@property (strong, nonatomic) UIWindow * window;
+
+@end