+++ Demo/Application/Features/PayPal - Billing Agreement/BraintreeDemoPayPalBillingAgreementViewController.m
@@ -25,6 +25,7 @@
     BTPayPalDriver *driver = [[BTPayPalDriver alloc] initWithAPIClient:self.apiClient];
     driver.appSwitchDelegate = self;
     BTPayPalRequest *checkout = [[BTPayPalRequest alloc] init];
+    checkout.activeWindow = self.view.window;
     [driver requestBillingAgreement:checkout completion:^(BTPayPalAccountNonce * _Nullable tokenizedPayPalCheckout, NSError * _Nullable error) {
         [sender setEnabled:YES];